replace libunwind's print stack trace
